% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles distinctes}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data frame ou tibble contenant une colonne \verb{Numéro de boucle}.}
}
\value{
Un entier correspondant au nombre de boucles distinctes.
}
\description{
Cette fonction calcule le nombre de valeurs distinctes présentes dans la colonne
\verb{Numéro de boucle} d'un jeu de données de trajets.
}
