% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data frame ou tibble contenant une colonne \code{Total}.}
}
\value{
Un nombre indiquant la somme totale des trajets.
}
\description{
Cette fonction additionne la colonne \code{Total} d'un jeu de données
contenant le nombre de trajets.
}
