% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_velo.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{World Health Organization TB data}
\format{
\subsection{\code{df_velo}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
df_velo
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
