% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les trajets avec anomalie potentielle}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data frame ou tibble contenant au moins les colonnes
\verb{Probabilité de présence d'anomalies} et \code{Total}.}
}
\value{
Un tibble filtré contenant uniquement les trajets répondant aux critères.
}
\description{
Cette fonction conserve les lignes où la probabilité de présence d'anomalies
est manquante, tout en gardant des valeurs de \code{Total} strictement positives
et inférieures à 10000.
}
