% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_trajet.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer un trajet selon un vecteur de numéros de boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant au moins la colonne
\verb{Numéro de boucle}.}

\item{boucle}{Un vecteur de numéros de boucle à sélectionner (numérique ou texte).}
}
\value{
Un tibble filtré contenant uniquement les boucles sélectionnées.
}
\description{
Cette fonction sélectionne les lignes du data.frame dont la colonne
\verb{Numéro de boucle} correspond aux valeurs fournies dans le vecteur \code{boucle}.
}
\examples{
filtrer_trajet(trajet = df_velo, boucle = c("880", "881"))
}
