% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet avec le total maximal}
\usage{
trouver_trajet_max(trajet)

trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data frame ou tibble contenant au moins les colonnes
\verb{Boucle de comptage}, \code{Jour} et \code{Total}, ainsi que celles nécessaires à
\code{filtre_anomalie()}.}
}
\value{
Un tibble avec les colonnes \verb{Boucle de comptage}, \code{Jour} et \code{Total}
correspondant au trajet maximal.

Un tibble avec les colonnes \verb{Boucle de comptage}, \code{Jour}, \code{Total},
\code{moyenne_jour_identique} et \code{moyenne_boucle_identique}.
}
\description{
Cette fonction renvoie la ou les lignes correspondant au trajet ayant
la valeur la plus élevée dans la colonne \code{Total}.

Cette fonction filtre d'abord les anomalies, puis identifie la ligne ayant
la valeur maximale de \code{Total}. Elle ajoute ensuite deux moyennes :
la moyenne des trajets pour le même jour et la moyenne des trajets pour
la même boucle de comptage.
}
